Checkpoint: Whole project

+++ src/util/sys.rs
@@ -11,17 +11,17 @@ pub fn get_cpu_count() -> usize {
 /// Retrieves the processor's timestamp counter (TSC) value. This is a high-resolution timer that counts the number of cycles since reset.
 #[cfg(target_arch = "x86_64")]
 #[inline(always)]
-pub fn get_cpu_rdtsc() -> u32 {
-    unsafe { std::arch::x86_64::_rdtsc() as u32 }
+pub fn get_cpu_rdtsc() -> u64 {
+    unsafe { std::arch::x86_64::_rdtsc() }
 }
 
 /// Fallback: Uses system time in nanoseconds.
 #[cfg(not(target_arch = "x86_64"))]
 #[inline(always)]
-pub fn get_cpu_rdtsc() -> u32 {
+pub fn get_cpu_rdtsc() -> u64 {
     use std::time::{SystemTime, UNIX_EPOCH};
     SystemTime::now()
         .duration_since(UNIX_EPOCH)
         .unwrap_or_default()
-        .subsec_nanos()
+        .subsec_nanos() as u64
 }
